AUT-1208 - Simplify code by removing verify-mfa-code-controller and moving logic into relevant controller
- We had bits of code where one controller was calling another controller in order to reduce code duplication. This causes some confusion and makes the code harder to read. We're not duplicating a great deal of code so in order to simplify it we can move the logic into the relevant controller.
